+++ llama_index/indices/knowledge_graph/retrievers.py
@@ -11,13 +11,11 @@ from llama_index.indices.knowledge_graph.base import KnowledgeGraphIndex
 from llama_index.indices.query.embedding_utils import get_top_k_embeddings
 from llama_index.indices.query.schema import QueryBundle
 from llama_index.indices.service_context import ServiceContext
-from llama_index.prompts.base import Prompt, PromptType
+from llama_index.prompts import BasePromptTemplate, PromptTemplate, PromptType
 from llama_index.prompts.default_prompts import DEFAULT_QUERY_KEYWORD_EXTRACT_TEMPLATE
-from llama_index.prompts.prompts import QueryKeywordExtractPrompt
 from llama_index.schema import BaseNode, MetadataMode, NodeWithScore, TextNode
 from llama_index.storage.storage_context import StorageContext
 from llama_index.utils import truncate_text
-
 
 DQKET = DEFAULT_QUERY_KEYWORD_EXTRACT_TEMPLATE
 DEFAULT_NODE_SCORE = 1000.0
@@ -54,9 +52,9 @@ class KGTableRetriever(BaseRetriever):
         query_keyword_extract_template (Optional[QueryKGExtractPrompt]): A Query
             KG Extraction
             Prompt (see :ref:`Prompt-Templates`).
-        refine_template (Optional[RefinePrompt]): A Refinement Prompt
+        refine_template (Optional[BasePromptTemplate]): A Refinement Prompt
             (see :ref:`Prompt-Templates`).
-        text_qa_template (Optional[QuestionAnswerPrompt]): A Question Answering Prompt
+        text_qa_template (Optional[BasePromptTemplate]): A Question Answering Prompt
             (see :ref:`Prompt-Templates`).
         max_keywords_per_query (int): Maximum number of keywords to extract from query.
         num_chunks_per_query (int): Maximum number of text chunks to query.
@@ -78,7 +76,7 @@ class KGTableRetriever(BaseRetriever):
     def __init__(
         self,
         index: KnowledgeGraphIndex,
-        query_keyword_extract_template: Optional[QueryKeywordExtractPrompt] = None,
+        query_keyword_extract_template: Optional[BasePromptTemplate] = None,
         max_keywords_per_query: int = 10,
         num_chunks_per_query: int = 10,
         include_text: bool = True,
@@ -108,6 +106,7 @@ class KGTableRetriever(BaseRetriever):
         self.graph_store_query_depth = graph_store_query_depth
         self.use_global_node_triplets = use_global_node_triplets
         self.max_knowledge_sequence = max_knowledge_sequence
+        self._verbose = kwargs.get("verbose", False)
 
     def _get_keywords(self, query_str: str) -> List[str]:
         """Extract keywords."""
@@ -135,10 +134,10 @@ class KGTableRetriever(BaseRetriever):
         query_bundle: QueryBundle,
     ) -> List[NodeWithScore]:
         """Get nodes for response."""
-        logger.info(f"> Starting query: {query_bundle.query_str}")
         node_visited = set()
         keywords = self._get_keywords(query_bundle.query_str)
-        logger.info(f"> Query keywords: {keywords}")
+        if self._verbose:
+            print_text(f"Extraced keywords: {keywords}\n", color="green")
         rel_texts = []
         cur_rel_map = {}
         chunk_indices_count: Dict[str, int] = defaultdict(int)
@@ -179,8 +178,8 @@ class KGTableRetriever(BaseRetriever):
                     continue
                 rel_texts.extend(
                     [
-                        f"{sub} {rel_obj}"
-                        for sub, rel_objs in rel_map.items()
+                        str(rel_obj)
+                        for rel_objs in rel_map.values()
                         for rel_obj in rel_objs
                     ]
                 )
@@ -219,7 +218,7 @@ class KGTableRetriever(BaseRetriever):
                 for node_id in node_ids:
                     chunk_indices_count[node_id] += 1
         elif len(self._index_struct.embedding_dict) == 0:
-            logger.error(
+            logger.warn(
                 "Index was not constructed with embeddings, skipping embedding usage..."
             )
 
@@ -276,16 +275,26 @@ class KGTableRetriever(BaseRetriever):
         rel_initial_text = (
             f"The following are knowledge sequence in max depth"
             f" {self.graph_store_query_depth} "
-            f"in the form of "
-            f"`subject [predicate, object, predicate_next_hop, object_next_hop ...]`"
+            f"in the form of directed graph like:\n"
+            f"`subject -[predicate]->, object, <-[predicate_next_hop]-,"
+            f" object_next_hop ...`"
         )
         rel_info = [rel_initial_text] + rel_texts
         rel_node_info = {
             "kg_rel_texts": rel_texts,
             "kg_rel_map": cur_rel_map,
         }
+        rel_info_text = "\n".join(
+            [
+                str(item)
+                for sublist in rel_info
+                for item in (sublist if isinstance(sublist, list) else [sublist])
+            ]
+        )
+        if self._verbose:
+            print_text(f"KG context:\n{rel_info_text}\n", color="blue")
         rel_text_node = TextNode(
-            text="\n".join(rel_info),
+            text=rel_info_text,
             metadata=rel_node_info,
             excluded_embed_metadata_keys=["kg_rel_map", "kg_rel_texts"],
             excluded_llm_metadata_keys=["kg_rel_map", "kg_rel_texts"],
@@ -294,8 +303,6 @@ class KGTableRetriever(BaseRetriever):
         sorted_nodes_with_scores.append(
             NodeWithScore(node=rel_text_node, score=DEFAULT_NODE_SCORE)
         )
-        rel_info_text = "\n".join(rel_info)
-        logger.info(f"> Extracted relationships: {rel_info_text}")
 
         return sorted_nodes_with_scores
 
@@ -320,7 +327,7 @@ KEYWORDS: {question}
 ----
 """
 
-DEFAULT_SYNONYM_EXPAND_PROMPT = Prompt(
+DEFAULT_SYNONYM_EXPAND_PROMPT = PromptTemplate(
     DEFAULT_SYNONYM_EXPAND_TEMPLATE,
     prompt_type=PromptType.QUERY_KEYWORD_EXTRACT,
 )
@@ -336,7 +343,7 @@ class KnowledgeGraphRAGRetriever(BaseRetriever):
         service_context (Optional[ServiceContext]): A service context to use.
         storage_context (Optional[StorageContext]): A storage context to use.
         entity_extract_fn (Optional[Callable]): A function to extract entities.
-        entity_extract_template Optional[QueryKeywordExtractPrompt]): A Query Key Entity
+        entity_extract_template Optional[BasePromptTemplate]): A Query Key Entity
             Extraction Prompt (see :ref:`Prompt-Templates`).
         entity_extract_policy (Optional[str]): The entity extraction policy to use.
             default: "union"
@@ -368,10 +375,10 @@ class KnowledgeGraphRAGRetriever(BaseRetriever):
         service_context: Optional[ServiceContext] = None,
         storage_context: Optional[StorageContext] = None,
         entity_extract_fn: Optional[Callable] = None,
-        entity_extract_template: Optional[QueryKeywordExtractPrompt] = None,
+        entity_extract_template: Optional[BasePromptTemplate] = None,
         entity_extract_policy: Optional[str] = "union",
         synonym_expand_fn: Optional[Callable] = None,
-        synonym_expand_template: Optional[Prompt] = None,
+        synonym_expand_template: Optional[BasePromptTemplate] = None,
         synonym_expand_policy: Optional[str] = "union",
         max_entities: int = 5,
         max_synonyms: int = 5,
@@ -443,7 +450,7 @@ class KnowledgeGraphRAGRetriever(BaseRetriever):
         self,
         query_str: str,
         handle_fn: Optional[Callable],
-        handle_llm_prompt_template: Optional[Prompt],
+        handle_llm_prompt_template: Optional[BasePromptTemplate],
         cross_handle_policy: Optional[str] = "union",
         max_items: Optional[int] = 5,
         result_start_token: str = "KEYWORDS:",
@@ -493,7 +500,7 @@ class KnowledgeGraphRAGRetriever(BaseRetriever):
         self,
         query_str: str,
         handle_fn: Optional[Callable],
-        handle_llm_prompt_template: Optional[Prompt],
+        handle_llm_prompt_template: Optional[BasePromptTemplate],
         cross_handle_policy: Optional[str] = "union",
         max_items: Optional[int] = 5,
         result_start_token: str = "KEYWORDS:",
@@ -649,8 +656,9 @@ class KnowledgeGraphRAGRetriever(BaseRetriever):
         context_string = (
             f"The following are knowledge sequence in max depth"
             f" {self._graph_traversal_depth} "
-            f"in the form of "
-            f"`subject predicate, object, predicate_next_hop, object_next_hop ...`"
+            f"in the form of directed graph like:\n"
+            f"`subject -[predicate]->, object, <-[predicate_next_hop]-,"
+            f" object_next_hop ...`"
             f" extracted based on key entities as subject:\n"
             f"{_new_line_char.join(knowledge_sequence)}"
         )