+++ llama_index/indices/postprocessor/pii.py
@@ -6,7 +6,7 @@ from typing import List, Optional, Dict, Tuple, Callable
 from llama_index.indices.postprocessor.node import BasePydanticNodePostprocessor
 from llama_index.indices.query.schema import QueryBundle
 from llama_index.indices.service_context import ServiceContext
-from llama_index.prompts.prompts import QuestionAnswerPrompt
+from llama_index.prompts.base import PromptTemplate
 from llama_index.schema import NodeWithScore, MetadataMode
 
 
@@ -56,7 +56,7 @@ class PIINodePostprocessor(BasePydanticNodePostprocessor):
 
     def mask_pii(self, text: str) -> Tuple[str, Dict]:
         """Mask PII in text."""
-        pii_prompt = QuestionAnswerPrompt(self.pii_str_tmpl)
+        pii_prompt = PromptTemplate(self.pii_str_tmpl)
         # TODO: allow customization
         task_str = (
             "Mask out the PII, replace each PII with a tag, and return the text. "