+++ llama_index/node_parser/simple.py
@@ -1,6 +1,10 @@
 """Simple node parser."""
-from pydantic import Field
 from typing import List, Optional, Sequence
+
+try:
+    from pydantic.v1 import Field
+except ImportError:
+    from pydantic import Field
 
 from llama_index.callbacks.base import CallbackManager
 from llama_index.callbacks.schema import CBEventType, EventPayload