+++ llama_index/text_splitter/sentence_splitter.py
@@ -1,7 +1,11 @@
 """Sentence splitter."""
 from dataclasses import dataclass
-from pydantic import Field, PrivateAttr
 from typing import Callable, List, Optional
+
+try:
+    from pydantic.v1 import Field, PrivateAttr
+except ImportError:
+    from pydantic import Field, PrivateAttr
 
 from llama_index.callbacks.base import CallbackManager
 from llama_index.callbacks.schema import CBEventType, EventPayload